core/config.py
import os
from dotenv import load_dotenv

load_dotenv()

ODDS_API_KEY = os.environ.get("ODDS_API_KEY", "")
ODDS_BASE_URL = "https://api.the-odds-api.com/v4"
SPORT_KEY_NBA = "basketball_nba"
DEFAULT_REGIONS = "us"
DEFAULT_MARKETS = "spreads"