requirements.txt
streamlit>=1.32
pandas>=2.1
numpy>=1.26
requests>=2.31
python-dateutil>=2.9
nba_api>=1.4
pillow>=10.0
playwright>=1.40
python-dotenv>=1.0
Pillow>=10.0.0
pyarrow>=14.0

PyYAML>=6.0
